# Mr5-School-API-main/docs/openapi.yaml
openapi: 3.0.3
info:
  title: MR5 AI-Powered LMS API
  description: |
    RESTful API for the AI-Powered Learning Management System with Virtual Classroom capabilities.
    
    ## Authentication
    All protected endpoints require a JWT token in the Authorization header:
    `Authorization: Bearer <token>`
    
    ## Rate Limiting
    - Standard: 100 requests/minute
    - AI endpoints: 20 requests/minute
  version: 2.0.0
  contact:
    name: MR5 Support
    email: support@mr5.edu
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.mr5.edu
    description: Production server
...
paths:
...
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found

    put:
      tags:
        - Users
      summary: Update user
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # ============ COURSES ============
...
  /payments/create-checkout:
    post:
      tags:
        - Payments
      summary: Create Stripe checkout session
      operationId: createCheckoutSession
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                courseId:
                  type: string
                successUrl:
                  type: string
                cancelUrl:
                  type: string
              required:
                - courseId
      responses:
        '200':
          description: Checkout session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                  url:
                    type: string
  /payments/webhook:
    post:
      tags:
        - Payments
      summary: Stripe webhook handler
      operationId: handleStripeWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed

  # ============ AI SERVICES ============
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, teacher, student]
        avatar:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        avatar:
          type: string
        password:
          type: string
...